Extract shared findStyleTagIndexes helper for paste HTML parsing
Both cleanHtmlComments.ts (core) and getStyleMetadata.ts (plugins) had
nearly identical extractHtmlIndexes functions and matching <style> tag
constants. Consolidate them into a single findStyleTagIndexes utility
exported from roosterjs-content-model-dom.
