+++ packages/roosterjs-content-model-core/lib/command/paste/cleanHtmlComments.ts
@@ -1,42 +1,25 @@
+import { findStyleTagIndexes } from 'roosterjs-content-model-dom';
+
 const HtmlCommentStart = '\x3C!--';
 const HtmlCommentStart2 = '<!--';
 const HtmlCommentEnd = '-->';
-const styleTag = '<style';
-const styleClosingTag = '</style>';
-const nonWordCharacterRegex = /\W/;
 
 /**
  * @internal
  * Exported only for unit test
  */
 export function cleanHtmlComments(html: string) {
-    let { styleIndex, styleEndIndex } = extractHtmlIndexes(html);
+    let { styleIndex, styleEndIndex } = findStyleTagIndexes(html);
 
     while (styleIndex > -1) {
         html = removeCommentsFromHtml(html, HtmlCommentStart, styleEndIndex, styleIndex);
         html = removeCommentsFromHtml(html, HtmlCommentStart2, styleEndIndex, styleIndex);
         html = removeCommentsFromHtml(html, HtmlCommentEnd, styleEndIndex, styleIndex);
 
-        ({ styleIndex, styleEndIndex } = extractHtmlIndexes(html, styleEndIndex + 1));
+        ({ styleIndex, styleEndIndex } = findStyleTagIndexes(html, styleEndIndex + 1));
     }
 
     return html;
-}
-
-function extractHtmlIndexes(html: string, startIndex: number = 0) {
-    const htmlLowercase = html.toLowerCase();
-    let styleIndex = htmlLowercase.indexOf(styleTag, startIndex);
-    let currentIndex = styleIndex + styleTag.length;
-    let nextChar = html.substring(currentIndex, currentIndex + 1);
-
-    while (!nonWordCharacterRegex.test(nextChar) && styleIndex > -1) {
-        styleIndex = htmlLowercase.indexOf(styleTag, styleIndex + 1);
-        currentIndex = styleIndex + styleTag.length;
-        nextChar = html.substring(currentIndex, currentIndex + 1);
-    }
-
-    const styleEndIndex = htmlLowercase.indexOf(styleClosingTag, startIndex);
-    return { styleIndex, styleEndIndex };
 }
 
 function removeCommentsFromHtml(
