+++ packages/roosterjs-content-model-plugins/lib/paste/WordDesktop/getStyleMetadata.ts
@@ -1,39 +1,21 @@
-import { getObjectKeys } from 'roosterjs-content-model-dom';
+import { findStyleTagIndexes, getObjectKeys } from 'roosterjs-content-model-dom';
 import type { WordMetadata } from './WordMetadata';
 
 const FORMATING_REGEX = /[\n\t'{}"]+/g;
 const STYLE_TAG = '<style';
-const STYLE_TAG_END = '</style>';
-const nonWordCharacterRegex = /\W/;
 
 function extractStyleTagsFromHtml(htmlContent: string): string[] {
     const styles: string[] = [];
 
-    let { styleIndex, styleEndIndex } = extractHtmlIndexes(htmlContent);
+    let { styleIndex, styleEndIndex } = findStyleTagIndexes(htmlContent);
     while (styleIndex >= 0 && styleEndIndex >= 0) {
         const styleContent = htmlContent
             .substring(styleIndex + STYLE_TAG.length, styleEndIndex)
             .trim();
         styles.push(styleContent);
-        ({ styleIndex, styleEndIndex } = extractHtmlIndexes(htmlContent, styleEndIndex + 1));
+        ({ styleIndex, styleEndIndex } = findStyleTagIndexes(htmlContent, styleEndIndex + 1));
     }
     return styles;
-}
-
-function extractHtmlIndexes(html: string, startIndex: number = 0) {
-    const htmlLowercase = html.toLowerCase();
-    let styleIndex = htmlLowercase.indexOf(STYLE_TAG, startIndex);
-    let currentIndex = styleIndex + STYLE_TAG.length;
-    let nextChar = html.substring(currentIndex, currentIndex + 1);
-
-    while (!nonWordCharacterRegex.test(nextChar) && styleIndex > -1) {
-        styleIndex = htmlLowercase.indexOf(STYLE_TAG, styleIndex + 1);
-        currentIndex = styleIndex + STYLE_TAG.length;
-        nextChar = html.substring(currentIndex, currentIndex + 1);
-    }
-
-    const styleEndIndex = htmlLowercase.indexOf(STYLE_TAG_END, startIndex);
-    return { styleIndex, styleEndIndex };
 }
 
 /**
